requirements.txt
pandas>=1.5.0
numpy>=1.23.0
scipy>=1.9.0
scikit-learn>=1.1.0
pyyaml>=6.0
tushare>=1.2.89
matplotlib>=3.6.0
seaborn>=0.12.0
jupyter>=1.0.0
ipython>=8.0.0
pytest>=7.0.0
pytest-cov>=4.0.0
black>=22.0.0
flake8>=5.0.0
mypy>=0.990
